calm down some compilers / hint by Rainer

+++ can-monitor/src/decoder-can-cs2.c
@@ -247,9 +247,9 @@ void decode_cs2_channel_data(unsigned char *buffer, uint32_t uid, int kanal, int
 	a_messwert->name = calloc(1, strlen(p) + 1);
 	strcpy(a_messwert->name, p);
 	p = next_string(p);
-	a_messwert->min_bereich = atof(p);
+	a_messwert->min_bereich = (float) atof(p);
 	p = next_string(p);
-	a_messwert->max_bereich = atof(p);
+	a_messwert->max_bereich = (float) atof(p);
 	p = next_string(p);
 	a_messwert->einheit = calloc(1, strlen(p) + 1);
 	strcpy(a_messwert->einheit, p);
